Extract folder browser scanning module
